% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{recoverNetwork}
\alias{recoverNetwork}
\title{Recover Network}
\usage{
recoverNetwork(
  data,
  lambda,
  exoeffects = 1,
  docv = 0,
  timeweights = 1,
  eta = 0.05,
  Wder = NA,
  adaestmethod = "cvxr",
  dopostols = 0,
  Wfixed = -1
)
}
\arguments{
\item{data}{Panel data. See vignette for format and restriction.}

\item{lambda}{Penalization parameters in the format (p1,p1*,p2)}

\item{exoeffects}{Restricts gamma=0 if exoeffects=0 (optional, default is exoeffects=1)}

\item{timeweights}{Introduces time weights whcih is either 1 (default) or a vector with the same length as the number of time periods. This allows for reweighting the observations and estimate the version of the codes with the dynamic Wt.}

\item{eta}{Used in the computation of the initial conditions as explained in Step 4 in Appendix B.2. Default is 0.05.}

\item{Wder}{The derivative of the objective function with respect to the elements of Wij. This can be user-supplied to avoid the cost of computing the derivative. Computational gains for large N can be substantial. Default is NA.}

\item{adaestmethod}{Selection of estimation method for the adaptive step. Only supported version (default) is cvxr.}

\item{dopostols}{Equal to 1 if post-OLS is to be performed. Default is not using post-OLS, so dopostols=0, and final estimates are drawn from the unpenalized GMM.}

\item{Wfixed}{Used if part of the network is known. In this case, supply a NxN matrix with NA in the (i,j) elements that that are not known (and thus to be estimated), and with the known network intensity in the elements that are known. Supply Wfixed=-1 if network is entirely unknown and to be estimated.}
}
\value{
List with initial conditions, elastic net, adaptive elastic net, and unpenalized GMM steps. See vignette for format.
}
\description{
Package from paper "Identifying Network Ties from Panel Data: Theory and an Application to Tax Competition"
}
